chore: 清理 web-backup 目录并更新依赖
- 删除 Lingadmin-master/web-backup 目录下的所有备份文件
- 更新 EdgeCommon、Lingapi 的 go.mod 和 go.sum
- 更新前端配置文件和视图组件
- 添加测试文件和工具函数

+++ Lingapi-master/internal/setup/utils.go
@@ -21,7 +21,7 @@ func CompareVersion(version1 string, version2 string) int8 {
 		return 0
 	}
 
-	return stringutil.VersionCompare(fixVersion(version1), fixVersion(version2))
+	return int8(stringutil.VersionCompare(fixVersion(version1), fixVersion(version2)))
 }
 
 func fixVersion(version string) string {
